Factored out common shader code
